CompilationWorkflowHandler subscribes to SourceCompileSuccessEvent, SourcesAddedEvent (#382)
