+++ tests/Functional/Services/CompilationWorkflowHandlerTest.php
@@ -5,13 +5,18 @@ declare(strict_types=1);
 namespace App\Tests\Functional\Services;
 
 use App\Entity\TestConfiguration;
+use App\Event\SourceCompile\SourceCompileSuccessEvent;
+use App\Event\SourcesAddedEvent;
 use App\Message\CompileSource;
 use App\Services\CompilationWorkflowHandler;
 use App\Services\JobStore;
 use App\Tests\AbstractBaseFunctionalTest;
+use App\Tests\Mock\MockSuiteManifest;
 use App\Tests\Services\TestTestFactory;
+use Psr\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\Transport\InMemoryTransport;
+use Symfony\Contracts\EventDispatcher\Event;
 
 class CompilationWorkflowHandlerTest extends AbstractBaseFunctionalTest
 {
@@ -19,6 +24,7 @@ class CompilationWorkflowHandlerTest extends AbstractBaseFunctionalTest
     private JobStore $jobStore;
     private InMemoryTransport $messengerTransport;
     private TestTestFactory $testFactory;
+    private EventDispatcherInterface $eventDispatcher;
 
     protected function setUp(): void
     {
@@ -44,6 +50,12 @@ class CompilationWorkflowHandlerTest extends AbstractBaseFunctionalTest
         self::assertInstanceOf(TestTestFactory::class, $testFactory);
         if ($testFactory instanceof TestTestFactory) {
             $this->testFactory = $testFactory;
+        }
+
+        $eventDispatcher = self::$container->get(EventDispatcherInterface::class);
+        self::assertInstanceOf(EventDispatcherInterface::class, $eventDispatcher);
+        if ($eventDispatcher instanceof EventDispatcherInterface) {
+            $this->eventDispatcher = $eventDispatcher;
         }
     }
 
@@ -136,6 +148,42 @@ class CompilationWorkflowHandlerTest extends AbstractBaseFunctionalTest
                     );
                 },
                 'expectedQueuedMessage' => new CompileSource('Test/test2.yml'),
+            ],
+        ];
+    }
+
+    /**
+     * @dataProvider subscribesToEventsDataProvider
+     */
+    public function testSubscribesToEvents(Event $event)
+    {
+        $job = $this->jobStore->getJob();
+        $job->setSources([
+            'Test/test1.yml',
+            'Test/test2.yml',
+        ]);
+        $this->jobStore->store($job);
+
+        self::assertCount(0, $this->messengerTransport->get());
+
+        $this->eventDispatcher->dispatch($event);
+
+        self::assertCount(1, $this->messengerTransport->get());
+    }
+
+    public function subscribesToEventsDataProvider(): array
+    {
+        return [
+            SourceCompileSuccessEvent::class => [
+                'event' => new SourceCompileSuccessEvent(
+                    '/app/source/Test/test1.yml',
+                    (new MockSuiteManifest())
+                        ->withGetTestManifestsCall([])
+                        ->getMock()
+                ),
+            ],
+            SourcesAddedEvent::class => [
+                'event' => new SourcesAddedEvent(),
             ],
         ];
     }
